scraper/requirements.txt
playwright==1.42.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.1
